+++ projects/com.nvidia.grcuda.test/src/com/nvidia/grcuda/test/util/mock/GrCUDAStreamManagerMock.java
@@ -49,16 +49,11 @@ public class GrCUDAStreamManagerMock extends GrCUDAStreamManager {
     GrCUDAStreamManagerMock(CUDARuntime runtime,
                             RetrieveNewStreamPolicyEnum retrieveStreamPolicy,
                             RetrieveParentStreamPolicyEnum parentStreamPolicyEnum) {
-        super(runtime, retrieveStreamPolicy, parentStreamPolicyEnum);
-    }
-
-    GrCUDAStreamManagerMock(CUDARuntime runtime,
-                            RetrieveNewStreamPolicyEnum retrieveStreamPolicy) {
-        super(runtime, retrieveStreamPolicy, RetrieveParentStreamPolicyEnum.SAME_AS_PARENT);
+        super(runtime, retrieveStreamPolicy, parentStreamPolicyEnum, false);
     }
 
     GrCUDAStreamManagerMock(CUDARuntime runtime) {
-        super(runtime, RetrieveNewStreamPolicyEnum.ALWAYS_NEW, RetrieveParentStreamPolicyEnum.SAME_AS_PARENT);
+        super(runtime, RetrieveNewStreamPolicyEnum.ALWAYS_NEW, RetrieveParentStreamPolicyEnum.SAME_AS_PARENT, false);
     }
 
     int numStreams = 0;
@@ -71,7 +66,10 @@ public class GrCUDAStreamManagerMock extends GrCUDAStreamManager {
     }
 
     @Override
-    public void assignEvent(ExecutionDAG.DAGVertex vertex) { }
+    public void assignEventStart(ExecutionDAG.DAGVertex vertex) { }
+
+    @Override
+    public void assignEventStop(ExecutionDAG.DAGVertex vertex) { }
 
     @Override
     public void syncStream(CUDAStream stream) { }