+++ projects/com.nvidia.grcuda/src/com/nvidia/grcuda/GrCUDAContext.java
@@ -89,7 +89,7 @@ public final class GrCUDAContext {
     public GrCUDAContext(Env env) {
         this.env = env;
 
-        this.grCUDAOptionMap = GrCUDAOptionMap.getInstance(env.getOptions());
+        this.grCUDAOptionMap = new GrCUDAOptionMap(env.getOptions());
 
         // Retrieve the dependency computation policy;
         DependencyPolicyEnum dependencyPolicy = grCUDAOptionMap.getDependencyPolicy();
@@ -107,7 +107,7 @@ public final class GrCUDAContext {
         Boolean inputPrefetch = grCUDAOptionMap.isInputPrefetch();
 
         // Initialize the execution policy;
-        LOGGER.fine("using" + executionPolicy.toString() + " execution policy");
+        LOGGER.info("using" + executionPolicy.toString() + " execution policy");
         switch (executionPolicy) {
             case SYNC:
                 this.grCUDAExecutionContext = new SyncGrCUDAExecutionContext(this, env, dependencyPolicy, inputPrefetch ? PrefetcherEnum.SYNC : PrefetcherEnum.NONE);