+++ projects/com.nvidia.grcuda/src/com/nvidia/grcuda/GrCUDAOptions.java
@@ -86,4 +86,6 @@ public final class GrCUDAOptions {
     @Option(category = OptionCategory.USER, help = "Set the location of the TensorRT library.", stability = OptionStability.STABLE) //
     public static final OptionKey<String> TensorRTLibrary = new OptionKey<>(TensorRTRegistry.DEFAULT_LIBRARY);
 
+    @Option(category = OptionCategory.USER, help = "Log the kernels execution time.", stability = OptionStability.STABLE) //
+    public static final OptionKey<Boolean> TimeComputation = new OptionKey<>(GrCUDAOptionMap.DEFAULT_TIME_COMPUTATION);
 }