+++ projects/com.nvidia.grcuda/src/com/nvidia/grcuda/runtime/computation/KernelExecution.java
@@ -67,6 +67,16 @@ public class KernelExecution extends GrCUDAComputationalElement {
     }
 
     @Override
+    public void setExecutionTime(int deviceId, float time) {
+        this.configuredKernel.addExecutionTime(deviceId, time);
+    }
+
+    @Override
+    public float getExecutionTimeOnDevice(int deviceId){
+        return this.configuredKernel.getExecutionTimeOnDevice(deviceId);
+    }
+
+    @Override
     public Object execute() {
         grCUDAExecutionContext.getCudaRuntime().cuLaunchKernel(kernel, config, args, this.getStream());
         return NoneValue.get();
@@ -133,7 +143,7 @@ public class KernelExecution extends GrCUDAComputationalElement {
 //        return "KernelExecution(" + configuredKernel.toString() + "; args=[" +
 //                Arrays.stream(args.getOriginalArgs()).map(a -> Integer.toString(System.identityHashCode(a))).collect(Collectors.joining(", ")) +
 //                "]" + "; stream=" + this.getStream() + ")";
-        String event = this.getEvent().isPresent() ? Long.toString(this.getEvent().get().getEventNumber()) : "NULL";
+        String event = this.getEventStop().isPresent() ? Long.toString(this.getEventStop().get().getEventNumber()) : "NULL";
         return "kernel=" + kernel.getKernelName() + "; args=[" +
                 Arrays.stream(args.getOriginalArgs()).map(a -> Integer.toString(System.identityHashCode(a))).collect(Collectors.joining(", ")) +
                 "]" + "; stream=" + this.getStream().getStreamNumber() + "; event=" + event;